# docs/api/openapi.yml
openapi: 3.1.0
info:
  title: Stratum API
  description: |
    A code collaboration platform for the AI engineering era.

    This specification is generated from the route code (`src/index.ts` and
    `src/routes/*.ts`) and covers every `/api/*` route, the root health check,
    the GitHub webhook receiver, and the SCIM 2.0 provisioning endpoints under
    `/scim/v2`.

    Not covered here:
    - HTML page routes (`/`, `/p/*`, `/changes/*`, `/auth/*` — including the
      SSO sign-in flow at `/auth/sso/*` — `/dev-login`, `/ui.css`, ...).
    - The git smart-HTTP endpoints (`/@{namespace}/{slug}/info/refs`,
      `git-upload-pack`, `git-receive-pack`) — documented separately in
      `docs/adr/005`.

    Authentication: most endpoints accept a Stratum API token (user or agent)
    as a `Authorization: Bearer <token>` header. Browser requests may instead be
    authenticated by the `stratum_session` cookie. Read endpoints on projects
    with `visibility: public` also accept anonymous requests. Admin endpoints
    additionally accept an `X-Admin-API-Key` header. The `/scim/v2` endpoints
    accept only a per-connection SCIM bearer token (`stratum_scim_*`) and no
    other credential.

    Several form-friendly endpoints return a `302` redirect instead of JSON when
    the request body is form-encoded rather than `application/json`; this spec
    documents the JSON behaviour.
  version: 1.0.0
servers:
  - url: https://your-instance.workers.dev
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /api/projects:
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      description: |
        Creates a project (and its backing Artifacts git repo) in the caller's
        namespace, or in an organization's namespace when `org` is given (the
        caller needs org write access). Also accepts form-encoded bodies.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: 1-64 char alphanumeric slug
                visibility:
                  type: string
                  enum: [private, public]
                  default: private
                files:
                  type: object
                  additionalProperties: { type: string }
                  description: Initial file map (path -> contents). Defaults to a single `.gitkeep`, or starter files when `seed` is true.
                seed:
                  type: boolean
                  description: Seed the repo with default starter files
                org:
                  type: string
                  description: Organization slug to own the project
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                type: object
                required: [id, name, namespace, slug]
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  namespace: { type: string }
                  slug: { type: string }
                  path: { type: string }
                  remote: { type: string }
                  commit:
                    type: string
                    description: SHA of the initial commit
                  visibility: { type: string, enum: [private, public] }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Organization not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects in the caller's namespace
      description: Returns an empty list for unauthenticated callers. Results are filtered to projects the caller can read.
      responses:
        "200":
          description: Projects the caller can read
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items: { $ref: "#/components/schemas/Project" }
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/changes/{id}:
    parameters:
      - $ref: "#/components/parameters/ChangeId"
    get:
      operationId: getChange
      tags: [changes]
      summary: Get a change with its eval runs and cost summary
      security: [{ bearerAuth: [] }, {}]
      responses:
        "200":
          description: Change detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  change: { $ref: "#/components/schemas/Change" }
                  evalRuns:
                    type: array
                    items: { $ref: "#/components/schemas/EvalRun" }
                  costs:
                    type: array
                    items:
                      type: object
                      description: Per-kind cost summary for this change
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/orgs:
    post:
      operationId: createOrg
      tags: [orgs]
      summary: Create an organization
      description: The creator becomes an org admin. The slug must not collide with an existing username.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string, description: 1-64 char alphanumeric slug }
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  org: { $ref: "#/components/schemas/Org" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listOrgs
      tags: [orgs]
      summary: List organizations the caller belongs to
      responses:
        "200":
          description: Organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  orgs:
                    type: array
                    items: { $ref: "#/components/schemas/Org" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/agents:
    post:
      operationId: createAgent
      tags: [agents]
      summary: Register an agent and mint its token
      description: The agent token is returned once, on creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                model: { type: string }
                description: { type: string }
                promptHash: { type: string }
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent: { $ref: "#/components/schemas/Agent" }
                  token:
                    type: string
                    description: Plaintext agent token, shown once
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listAgents
      tags: [agents]
      summary: List the caller's agents
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items: { $ref: "#/components/schemas/Agent" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/bulk-import:
    post:
      operationId: startBulkImport
      tags: [bulk-import]
      summary: Start a bulk import job (up to 50 repositories)
      description: Imports run in the background; poll `GET /api/bulk-import/{id}` for status. Only the caller's own namespace is allowed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repos]
              properties:
                repos:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items:
                    type: object
                    required: [url]
                    properties:
                      url: { type: string }
                      namespace: { type: string }
                      slug: { type: string }
                      branch: { type: string }
                      visibility:
                        type: string
                        enum: [private, public]
      responses:
        "201":
          description: Job started
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobId: { type: string }
                  status: { type: string, example: queued }
                  totalRepos: { type: integer }
                  message: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: listBulkImportJobs
      tags: [bulk-import]
      summary: List the caller's bulk import jobs
      responses:
        "200":
          description: Jobs, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/BulkImportJobSummary" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/webhooks/github:
    post:
      operationId: receiveGitHubWebhook
      tags: [github]
      summary: GitHub webhook receiver
      description: |
        Inbound receiver for GitHub `push`, `pull_request`,
        `pull_request_review`, and `ping` events. Authenticated by HMAC
        signature (`X-Hub-Signature-256`) against the configured webhook
        secret, not by bearer token. Requires `X-GitHub-Event` and
        `X-GitHub-Delivery` headers; duplicate delivery ids are skipped.
        Processing errors still return 200 to prevent GitHub retries.
      security: []
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema: { type: string }
        - name: X-GitHub-Event
          in: header
          required: true
          schema: { type: string }
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw GitHub webhook payload for the event type
      responses:
        "200":
          description: Received (possibly a duplicate, or with a logged processing error)
          content:
            application/json:
              schema:
                type: object
                properties:
                  received: { type: boolean }
                  duplicate: { type: boolean }
                  error: { type: string }
        "400":
          description: Missing headers or invalid JSON
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Invalid signature
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "501":
          description: Webhook secret not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/admin/audit:
    get:
      operationId: adminAudit
      tags: [admin]
      summary: Query the audit trail
      security:
        - bearerAuth: []
        - adminApiKey: []
      parameters:
        - name: action
          in: query
          schema: { type: string }
          description: Filter by audit action (e.g. merge.forced, token.rotated)
        - name: actor
          in: query
          schema: { type: string }
          description: Filter by actor id
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEntry" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/admin/backup:
    get:
      operationId: adminListBackups
      tags: [admin]
      summary: List backup runs (newest first)
      security:
        - bearerAuth: []
        - adminApiKey: []
      responses:
        "200":
          description: Backup runs, each flagged complete/incomplete
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs:
                    type: array
                    items: { type: object }
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: adminRunBackup
      tags: [admin]
      summary: Trigger a backup run now (single-flight)
      security:
        - bearerAuth: []
        - adminApiKey: []
      responses:
        "200":
          description: Backup summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary: { type: object }
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A backup run is already in progress
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          $ref: "#/components/responses/InternalError"
...
  /scim/v2/Users:
    get:
      operationId: scimListUsers
      tags: [scim]
      summary: List users in the connection's scope
      security:
        - scimBearer: []
      parameters:
        - name: filter
          in: query
          schema: { type: string }
          description: Only `userName eq "..."` and `externalId eq "..."` are supported; any other filter answers 501
        - name: startIndex
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
          description: 1-based start index
        - name: count
          in: query
          schema: { type: integer, minimum: 0, maximum: 200, default: 100 }
          description: Page size; 0 returns totalResults with no Resources
      responses:
        "200":
          description: SCIM ListResponse of Users
          content:
            application/scim+json:
              schema: { $ref: "#/components/schemas/ScimUserList" }
        "401":
          $ref: "#/components/responses/ScimUnauthorized"
        "501":
          description: Unsupported filter expression
          content:
            application/scim+json:
              schema: { $ref: "#/components/schemas/ScimError" }
    post:
      operationId: scimCreateUser
      tags: [scim]
      summary: Provision a user
      description: |
        `userName` must be an email within the connection's verified domains.
        If an account with that email already exists (and is not being
        deleted), it is adopted into the connection's scope and answered with
        200 instead of 201; otherwise a new account is created and added to
        the org as a member.
      security:
        - scimBearer: []
      requestBody:
        required: true
        content:
          application/scim+json:
            schema:
              type: object
              required: [userName]
              properties:
                userName: { type: string, description: Email address within the connection's verified domains }
                externalId: { type: string }
                active: { type: boolean, default: true }
      responses:
        "200":
          description: Existing account adopted
          content:
            application/scim+json:
              schema: { $ref: "#/components/schemas/ScimUser" }
        "201":
          description: User created
          content:
            application/scim+json:
              schema: { $ref: "#/components/schemas/ScimUser" }
        "400":
          description: Invalid attribute (SCIM error body)
          content:
            application/scim+json:
              schema: { $ref: "#/components/schemas/ScimError" }
        "401":
          $ref: "#/components/responses/ScimUnauthorized"
        "409":
          description: Conflicting account, already-provisioned user, or duplicate externalId
          content:
            application/scim+json:
              schema: { $ref: "#/components/schemas/ScimError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Stratum API token (user token or agent token)
    adminApiKey:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: Admin API key for administrator endpoints
    scimBearer:
      type: http
      scheme: bearer
      description: >
        Per-connection SCIM token (`stratum_scim_*`), rotated via
        `POST /api/orgs/{slug}/sso/scim-token`. The only credential the
        `/scim/v2` endpoints accept — session cookies and user/agent tokens
        are never honored there.
  parameters:
...
    ChangeId:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: Change id
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found (also returned for access denied, to avoid disclosing existence)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: >
            Machine-readable code present on some errors (e.g. TARGET_DELETING,
            STALE_BASE, STALE_WORKSPACE, WORKSPACE_UNVERIFIABLE, MERGE_CONFLICT,
            PROTECTION_BLOCKED, NOT_REDRIVABLE, GONE, INVALID_PATH,
            DOMAIN_VERIFICATION_FAILED)
...
    ScimUserList:
      type: object
      description: SCIM ListResponse envelope
      properties:
        schemas:
          type: array
          items: { type: string }
        totalResults: { type: integer }
        startIndex: { type: integer }
        itemsPerPage: { type: integer }
        Resources:
          type: array
          items: { $ref: "#/components/schemas/ScimUser" }
...
    Project:
      type: object
      required: [id, name, namespace, slug]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        namespace:
          type: string
          description: '"@username" or "@org-slug"'
        slug: { type: string }
        path:
          type: string
          description: "/{namespace}/{slug}"
        remote:
          type: string
          description: Backing Artifacts git remote URL
        createdAt: { type: string, format: date-time }
        visibility:
          type: string
          enum: [private, public]
        githubUrl: { type: string }
        githubOwner: { type: string }
        githubRepo: { type: string }
        githubDefaultBranch: { type: string }
        githubConnectionStatus:
          type: string
          enum: [connected, disconnected]
...
    Change:
      type: object
      required: [id, project, workspace, status, createdAt]
      properties:
        id: { type: string }
        project:
          type: string
          description: Project name (or project id for changes created via the GitHub bridge)
        projectId:
          type: string
          format: uuid
          description: Globally-unique project id; absent on legacy rows
        workspace: { type: string }
        status:
          type: string
          enum: [open, needs_changes, accepted, approved, promoted, merged, rejected, reverted]
        agentId: { type: string }
        evalScore: { type: number }
        evalPassed: { type: boolean }
        evalReason: { type: string }
        baseSha:
          type: string
          description: Project HEAD at change creation — the base the evaluation ran against
        evaluatedSha:
          type: string
          description: Workspace tip the evaluation ran against; merges reject if it moved
        evaluatedTreeOid: { type: string }
        agentModel:
          type: string
          description: Authoring agent's model, snapshotted at change creation
        agentPromptHash: { type: string }
        workspaceHeadSha: { type: string }
        createdAt: { type: string, format: date-time }
        mergedAt: { type: string, format: date-time }
        githubOwner: { type: string }
        githubRepo: { type: string }
        githubBranch: { type: string }
        githubPrNumber: { type: integer }
        githubPrUrl: { type: string }
        githubPrState: { type: string }
        githubHeadSha: { type: string }
        promotedAt: { type: string, format: date-time }
        promotedBy: { type: string }
...
    EvalRun:
      type: object
      required: [id, changeId, evaluatorType, score, passed, reason, ranAt]
      properties:
        id: { type: string }
        changeId: { type: string }
        evaluatorType:
          type: string
          description: e.g. secret_scan, diff, webhook, llm, sandbox
        score: { type: number }
        passed: { type: boolean }
        reason: { type: string }
        issues:
          type: array
          items: { type: string }
        ranAt: { type: string, format: date-time }
...
    Org:
      type: object
      required: [id, name, slug, ownerId, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        ownerId: { type: string }
        createdAt: { type: string, format: date-time }
...
    Agent:
      type: object
      required: [id, name, ownerId, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        ownerId: { type: string }
        model: { type: string }
        description: { type: string }
        promptHash: { type: string }
        createdAt: { type: string, format: date-time }
...
    BulkImportJobSummary:
      type: object
      properties:
        jobId: { type: string }
        status:
          type: string
          enum: [queued, processing, completed, failed, partial]
        totalRepos: { type: integer }
        processedRepos: { type: integer }
        successfulRepos: { type: integer }
        failedRepos: { type: integer }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        hasErrors: { type: boolean }
...